feat(app): M16 — error states, brand identity, M15b cleanup (#38)
* feat(app): M16 — differentiated connect-failure copy (Farsi + English)

The main screen's single generic "Connection failed" copy hid which of
four distinct onboarding-adjacent failures was actually happening. Now
the user sees the specific reason and what to do about it:

  NO_INTERNET    → "No internet — check your connection"
  VPN_REVOKED    → "VPN permission revoked — re-grant in Android settings"
  NO_ACCESS      → "No access URL saved — import one first"
  SIDECAR_FAILED → "Core failed to start — try again"
  UNKNOWN        → existing generic fallback

Implementation:
- ParvazVpnService.FailReason enum + failReason field on SessionState.
- Each fail() call site picks its reason (no Access → NO_ACCESS, null
  establish() → VPN_REVOKED, CoreLauncher.start() failure → SIDECAR_FAILED).
- Added hasInternet() via ConnectivityManager at the top of
  scheduleStart() — NET_CAPABILITY_INTERNET check, fails fast before
  opening the TUN so the user doesn't see a pointless connect-then-fail.
- ACCESS_NETWORK_STATE added to the manifest (normal permission, no
  runtime prompt).
- MainUiState carries failReason, MainScreen.DisconnectedStamp picks
  the right string resource via failReasonStringRes().

Farsi copy is the default; English lives in values-en/. Server-
unreachable distinction isn't wired yet (no liveness probe against
Apps Script at connect time); future follow-up.

Clears 4 of 5 PLAN.md M16 cases.


* fix(app): require VALIDATED network before VPN start (codex M16 review)

INTERNET is an operator bit — a Wi-Fi hotspot behind a captive portal
or a router with a downed upstream can still advertise it. The
pre-connect check would then pass, the TUN would open, traffic would
fail, and the user would get a generic "Connection failed" instead of
"No internet".

VALIDATED is the signal that Android's NetworkValidator ran its
HTTP/204 probe and got a 2xx — actual end-to-end reachability.
Requiring both closes the captive-portal gap.


* docs(superpowers): integration spec for M15b → main → m16 rebase

Locks the path for landing tun2socks + DNS work that's been sitting
on feature branches since M15b-alpha/beta. Path 1: PR feat/m15b-beta-dns
into main, then rebase m16-error-states with conflict-resolution rules
that preserve M16's FailReason granularity over M15b's plain fail().


* docs(superpowers): pin FailReason placement after rebase

Adds ParvazConnectionState.kt row to the conflict table — that file is
new in M15b-beta and FailReason should move into it alongside
ConnectionState + SessionState rather than staying nested in
ParvazVpnService. Also adds MainScreen.kt row (M16's failReasonStringRes
lookup needs the matching import-path tweak).


* wip(m16): park brand-identity drawables + codex-review fixes pre-rebase

Snapshot of dirty working tree from m16-error-states before rebasing
onto M15b-on-main. Two distinct bodies of work parked here:

1. Brand identity WIP — adaptive launcher + cold-launch SplashScreen
   API (Theme.Parvaz.Starting + ic_splash_mark + parvaz_mark) +
   ic_launcher_monochrome + docs/identity{.md,/parvaz-mark.svg} +
   website parity (favicon.svg, parvaz-mark.svg). Replaces the AGP
   Bugdroid raster across mipmap-*dpi.

2. M16 codex-review fixes — INTERNET permission (production bug, not
   just a test fix), CoreLauncher coroutine drainer + @Volatile,
   CaInstallScreen UNVERIFIED state for Samsung KeyChain hide,
   instrumentation test Byte/Int assertEquals fix, PLAN.md trim
   under the 200-line CI rule, MainActivity SplashScreen integration.

To replay onto rebased m16-error-states: cherry-pick this commit, then
resolve any collisions with M15b's themes.xml / AndroidManifest changes.


* fix(app): trim ParvazVpnService for the 200-line CI cap

After the M15b rebase + WIP replay the file landed at 219 lines
(M15b plumbing + M16 FailReason granularity + brand-WIP imports).
Compress the class doc and three in-line comments — all WHYs kept,
just terser. Lands the file at exactly 200 lines.


---------

+++ app/src/androidTest/java/dk/cocode/parvaz/mitm/CaInstallerInstrumentedTest.kt
@@ -38,7 +38,7 @@ class CaInstallerInstrumentedTest {
         val name = intent.getStringExtra(KeyChain.EXTRA_NAME)
         assertNotNull(extra)
         assertEquals(5, extra!!.size)
-        assertEquals(1, extra[0])
+        assertEquals(1.toByte(), extra[0])
         assertEquals("Parvaz Test CA", name)
     }
